easygo requirements.txt
flask
paddleocr
opencv-python-headless
pillow
requests